# Conflicts:
#	src/main/java/ar/edu/utn/frba/dds/models/repositories/heladera/MovimientosHeladeraRepository.java
#	src/test/java/ar/edu/utn/frba/dds/models/repositories/heladera/MovimientosHeladeraRepositoryTest.java

+++ src/main/java/ar/edu/utn/frba/dds/models/repositories/contribucion/EntregaTarjetasRepository.java
@@ -24,7 +24,7 @@ public class EntregaTarjetasRepository {
     return instancia;
   }
 
-  public Optional<EntregaTarjetas> get(int id) {
+  public Optional<EntregaTarjetas> get(Long id) {
     return entregas.stream().filter(entrega -> entrega.getId() == id).findFirst();
   }
 
@@ -36,7 +36,7 @@ public class EntregaTarjetasRepository {
         .sum();
   }
 
-  public int insert(EntregaTarjetas entrega) throws RepositoryException {
+  public Long insert(EntregaTarjetas entrega) throws RepositoryException {
     if (entregas
         .stream()
         .flatMap(entregaAnterior -> entregaAnterior.getTarjetasRepartidas().stream())
@@ -45,7 +45,7 @@ public class EntregaTarjetasRepository {
     }
 
     entregas.add(entrega);
-    entrega.setId(entregas.size());
+    entrega.setId((long) entregas.size());
 
     return entrega.getId();
   }
